Seperate sw update state to download and delivery
Sota download is split in to two stages. Sota packages are stored
in a temporary file first after which the object 9 state is set
to DOWNLOADED. On the next stage the downloaded package is read
by the update daemon and unpacked. The object9 state is set to
DELIVERED at the end of stage2.

Use semaphore to co-ordinate user agreement.

Resolves: LE-6032
Change-Id: If8d155de9e35deb8785ab332341a1fb09bf02969