Potential crash when a connection is stopped
When a connection to the server results in a connection failure to the
Device Management server, a connection is initiated to the bootstrap
server. Then a connection is initiated to the Device Management server.
When the connection is stopped, a potential crash could happen because
list of connection was not correctly released.

Resolves: LE-6580
Change-Id: I3c901d2064fc74299ccde3858e2ffbe088f2ffa0