Ask for user agreement on download resume
This patch changes the behaviour of the package downloader.
In the previous version once the user agrees with the download all the
other actions will be token automatically. For example the package
downloader resumes the download automatically when a session starts.
In this version, the package downloader will ask for user agreement each
time it tries to resume a download.

Resolves: LE-7871
Change-Id: I875ee0a792f65865aca5344204de31025dc6508b