Fix bootstrap configuration loading issue
Issue:
Due to a parameter added in the bootstrap configuration file,
updating Legato to a recent version may impact the file decoding.
If the bootstrap configuration file fails, then the AVC polling
timer can't be configured until a manual connection to the
bootstrap server is made. Also, the old lifetime value may be lost.

Fix:
Detect the parameter presence in the configuration file

Resolves: LE-11324
Change-Id: I60f0a82364161c7f16a580535c5011cba6adabaa