[FOTA] Check for firmware update state before resuming download
Issue:
In case of user agreement enabled scenario, the module goes into
reboot before the user accepts the download. When the user accepts
the download after the device boots up, the download is triggered
to resume. The position to resume the download starts from a random
position instead of starting from 0. Hence, the download fails as
eventually the total download count exceeds the the full image length.
There is no method known to reset the resume position given by the
modem.

Solution:
The package downloader workspace package size is already the full
package size before the user accepts download. Add an additional
condition to check the firmware update state to decide whether to
resume the download or not. If the firmware update state is at wait
download or idle, download is not resumed.

Resolves: LE-14379
Change-Id: Id74a972d7c2c3ecc976ebda0786fb217ab25aac9