Do not connect to every server when bootstrapping
When a bootstrap occurs, we get keys for every server (AVC and EDM if enabled).This will then trigger the device to connect to every server using the DM keys. However, when we initiate a connection using AT+WDSS=1,1 and it bootstraps, we should just connect to AVC to keep the behaviour the same. Otherwise, if the SIM is not compatible with EDM, a basic connection with AVC will fail.

Resolves: LE-15650
Change-Id: I60abf2c6a03438d7df590784c297f352ee028714