Check tmp before dereferencing it
On initial startup the connection may not yet be in the DTLS connection
list.  In this case attempting to remove the connection would crash as
the loop would hit the end of the list and dereference a NULL pointer.

Check the iterator pointer is not NULL before dereferencing it, and
break out of the loop if it is.  Ironically, the case where the pointer
is NULL on exit from the loop is already handled.

This resolves the crash in LE-16390 but not the failure to connect.

Resolves: LE-16390
Change-Id: If69cb149f4591721bd31d7ecd1e8d5e596fcb543